Realizados cambios en varias clases y añadida las relaciones de las clases conductor camion y paquete

+++ src/main/java/com/josefco/accesoadatosaa/repository/CamionRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CamionRepository extends CrudRepository<Camion, Long> {
+public interface CamionRepository extends CrudRepository<Camion, Integer> {
 
     List<Camion> findAll();
     Camion findById(int id);
